+++ src/main/java/com/example/internintelligence_potfolioapidevelopment/controller/SkillController.java
@@ -1,6 +1,5 @@
 package com.example.internintelligence_potfolioapidevelopment.controller;
 
-import com.example.internintelligence_potfolioapidevelopment.dto.ExperienceDto;
 import com.example.internintelligence_potfolioapidevelopment.dto.SkillDto;
 import com.example.internintelligence_potfolioapidevelopment.service.SkillService;
 import io.swagger.v3.oas.annotations.Operation;
@@ -21,7 +20,7 @@ public class SkillController {
 
     @Operation(summary = "Retrieves another user's skills list")
     @GetMapping("/users/{userId}")
-    public List<SkillDto> getAllSkillsByUserId(@PathVariable Long userId){
+    public List<SkillDto> getAllSkillsByUserId(@PathVariable Integer userId){
         return skillService.getAllSkillsByUserId(userId);
     }
 
@@ -38,13 +37,13 @@ public class SkillController {
     }
 
     @PutMapping("/edit/{skillId}")
-    public SkillDto editExperience(HttpServletRequest request, @PathVariable Long skillId, @Validated @RequestBody SkillDto skillDto){
+    public SkillDto editExperience(HttpServletRequest request, @PathVariable Integer skillId, @Validated @RequestBody SkillDto skillDto){
         return skillService.editSkill(request, skillId,skillDto);
     }
 
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @DeleteMapping("/delete/{skillId}")
-    public void deleteExperience(HttpServletRequest request,@PathVariable Long skillId){
+    public void deleteExperience(HttpServletRequest request,@PathVariable Integer skillId){
         skillService.deleteSkill(request,skillId);
     }
